05.04.24.05.04
Fix #9

API now fetches raw system data during faction request. This includes system's XYZ coordinates, but does not include influence figures.

Influence figures can be obtained from the /systems endpoint using the factionDetails parameter. However, fetching this seems to be rate limited in some way.

So instead, we get raw system data from the /faction endpoint using the systemDetails param, and use that data as it is. Once a system has been placed in trackedSystems, a check is run to find which systems need influence figures, and a second API call set is run in the background to obtain this, and updates the tracked systems with new influence figures.

This results in immediate mapping and display of the systems, while the slower collection of influence figures can be run in the background.

A status indicator has been added in the lower left of the screen to indicate that these background activities are occurring. Systems will show "LOADING" until the influence for the given system has been updated.